backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
pillow
psycopg2-binary
PyJWT
sqlparse
tzdata